--- src/frontend/App/typer/brev.ts
@@ -1,12 +1,0 @@
-export interface IAvsnitt {
-    deloverskrift: string;
-    innhold: string;
-    id: string;
-}
-
-export interface IFritekstBrev {
-    overskrift: string;
-    avsnitt: IAvsnitt[];
-    fagsakId?: string;
-    behandlingId?: string;
-}
